add useImageUpload hooks
