+++ client/components/images/index.ts
@@ -1,2 +1,1 @@
 export { default as AvatarImage } from './AvatarImage';
-export { default as ImageUpload } from './ImageUpload';
